2012 ...
second commit

third initial commit

first working commit

first working commit

split out work exports & aliases from the other stuff

fix history

clean up

bash prompt

Initial commit

initial gh commit

add liquid highlighting

Revert "add liquid highlighting"

This reverts commit 7481f9c9358d15e8190c415047cbf5a2064b9562.

fix prompt

file conditinal

tweak prompt

filetype + vim liquid

updates & plugins

enhance readme

fix gatherconf

zen-coding & a jksky alias

add futzme function

buildguild

tweak buildguild alias

tweaks

tweaks

add some aliases, tweak vim colors

vim tweaks

tweak vim

fix char highlighting

fix tabs in vim

git ignore

+++ make.sh
@@ -8,7 +8,7 @@
 
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old             # old dotfiles backup directory
-files="bash_profile vimrc vim"    # list of files/folders to symlink in homedir
+files="bashrc bash_profile vimrc vim"    # list of files/folders to symlink in homedir
 
 ##########
 
@@ -24,8 +24,10 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [[ -f $file ]]; then
+        echo "Moving any existing dotfiles from ~ to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+        echo "Creating symlink to $file in home directory."
+        ln -s $dir/$file ~/.$file
+    fi
 done
